+++ src/Console/Command.php
@@ -35,10 +35,9 @@ class Command extends BaseCommand
     public function execute(Arguments $args, ConsoleIo $io): ?int
     {
         $config = $this->getConnection()->config();
-        $driver = $this->getDriver($this->getConnection());
 
         $io->out(__d('database_backup', 'Connection: {0}', $config['name']));
-        $io->out(__d('database_backup', 'Driver: {0}', get_class_short_name($driver)));
+        $io->out(__d('database_backup', 'Driver: {0}', $this->getDriverName($this->getConnection())));
         $io->hr();
 
         return null;